check if user is admin / can patch

+++ api/applications/applications_handler.go
@@ -218,7 +218,16 @@ func (ah *ApplicationHandler) ChangeRegistrationDetails(ctx context.Context, app
 		return nil, radixhttp.ValidationError("Radix Registration", fmt.Sprintf("App name %s does not correspond with application name %s", appName, application.Name))
 	}
 
-	// Make check that this is an existing application
+	// Make check that this is an existing application and that the user has access to it
+	userAccount := ah.getUserAccount()
+	userIsAdmin, err := utils.UserIsAdmin(ctx, &userAccount, application.Name)
+	if err != nil {
+		return nil, err
+	}
+	if !userIsAdmin {
+		return nil, fmt.Errorf("user is not allowed to change registration details for %s", appName)
+	}
+
 	currentRegistration, err := ah.getUserAccount().RadixClient.RadixV1().RadixRegistrations().Get(ctx, appName, metav1.GetOptions{})
 	if err != nil {
 		return nil, err
@@ -266,7 +275,16 @@ func (ah *ApplicationHandler) ChangeRegistrationDetails(ctx context.Context, app
 
 // ModifyRegistrationDetails handler for ModifyRegistrationDetails
 func (ah *ApplicationHandler) ModifyRegistrationDetails(ctx context.Context, appName string, applicationRegistrationPatchRequest applicationModels.ApplicationRegistrationPatchRequest) (*applicationModels.ApplicationRegistrationUpsertResponse, error) {
-	// Make check that this is an existing application
+	// Make check that this is an existing application and that the user has access to it
+	userAccount := ah.getUserAccount()
+	userIsAdmin, err := utils.UserIsAdmin(ctx, &userAccount, appName)
+	if err != nil {
+		return nil, err
+	}
+	if !userIsAdmin {
+		return nil, fmt.Errorf("user is not allowed to modify registration details for %s", appName)
+	}
+
 	currentRegistration, err := ah.getUserAccount().RadixClient.RadixV1().RadixRegistrations().Get(ctx, appName, metav1.GetOptions{})
 	if err != nil {
 		return nil, err
@@ -381,10 +399,14 @@ func (ah *ApplicationHandler) ModifyRegistrationDetails(ctx context.Context, app
 
 // DeleteApplication handler for DeleteApplication
 func (ah *ApplicationHandler) DeleteApplication(ctx context.Context, appName string) error {
-	// Make check that this is an existing application
-	_, err := ah.getUserAccount().RadixClient.RadixV1().RadixRegistrations().Get(ctx, appName, metav1.GetOptions{})
+	// Make check that this is an existing application and that the user has access to it
+	userAccount := ah.getUserAccount()
+	userIsAdmin, err := utils.UserIsAdmin(ctx, &userAccount, appName)
 	if err != nil {
 		return err
+	}
+	if !userIsAdmin {
+		return fmt.Errorf("user is not allowed to delete registration %s", appName)
 	}
 
 	err = ah.getUserAccount().RadixClient.RadixV1().RadixRegistrations().Delete(ctx, appName, metav1.DeleteOptions{})
@@ -510,7 +532,8 @@ func (ah *ApplicationHandler) triggerPipelineBuildOrBuildDeploy(ctx context.Cont
 
 	// Check if branch is mapped
 	if !applicationconfig.IsConfigBranch(branch, radixRegistration) {
-		application, err := utils.CreateApplicationConfig(ctx, ah.getUserAccount().Client, ah.getUserAccount().RadixClient, ah.getUserAccount().SecretProviderClient, appName)
+		userAccount := ah.getUserAccount()
+		application, err := utils.CreateApplicationConfig(ctx, &userAccount, appName)
 		if err != nil {
 			return nil, err
 		}
@@ -626,7 +649,16 @@ func (ah *ApplicationHandler) getMachineUserServiceAccount(ctx context.Context, 
 
 // RegenerateDeployKey Regenerates deploy key and secret and returns the new key
 func (ah *ApplicationHandler) RegenerateDeployKey(ctx context.Context, appName string, regenerateDeployKeyAndSecretData applicationModels.RegenerateDeployKeyAndSecretData) error {
-	// Make check that this is an existing application and user has access to it
+	// Make check that this is an existing application and that the user has access to it
+	userAccount := ah.getUserAccount()
+	userIsAdmin, err := utils.UserIsAdmin(ctx, &userAccount, appName)
+	if err != nil {
+		return err
+	}
+	if !userIsAdmin {
+		return fmt.Errorf("user is not allowed to regenerate deploy key for %s", appName)
+	}
+
 	currentRegistration, err := ah.getUserAccount().RadixClient.RadixV1().RadixRegistrations().Get(ctx, appName, metav1.GetOptions{})
 	if err != nil {
 		return err
